Make dotprod verifiable with modern versions of SAW (#6)
Cryptol 2.9.0+ (which SAW 0.6+ bundles) removes the `Arith` class in favor of
`Ring`. Furthermore, Cryptol 2.11.0+ (which SAW 0.8+ bundles) adds an `Eq`
constraint on `sum`. To allow SAW 0.6 or later to verify the `dotprod` example,
which uses `Arith` and `sum`, I removed the `(Arith a)` constraint in favor of
`(Eq a, Ring a)`. This means that the examples now require SAW 0.6+, so I added
a mention of this in the `README`.